backend/requirements.txt
# Core dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
Pillow==10.0.0
Werkzeug==3.0.0
SQLAlchemy==2.0.0
psutil==5.9.8
psycopg2-binary==2.9.9

# Production dependencies
gunicorn==21.2.0
supervisor==4.2.5
python-dotenv==1.0.0

# Testing dependencies
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0
coverage==7.4.1

# Development dependencies
black==24.2.0
flake8==7.0.0
mypy==1.8.0